backend/requirements.txt
flask==3.0.2
pymongo==4.6.2
pyjwt==2.8.0
flask-cors==4.0.0
pypdf==4.0.1
spacy==3.7.4
transformers==4.38.1
werkzeug==3.0.1
python-dotenv==1.0.1
